abstract bash cmd running to pkg folder
