+++ cli/cmd/start/backend/backend.go
@@ -5,15 +5,12 @@ package backend
 
 import (
 	"fmt"
-	"io"
-	"os"
 	"os/exec"
 	"strings"
 
 	"github.com/DSGT-DLP/Deep-Learning-Playground/cli/cmd/start"
-	"github.com/creack/pty"
+	"github.com/DSGT-DLP/Deep-Learning-Playground/cli/pkg"
 	"github.com/spf13/cobra"
-	"golang.org/x/term"
 )
 
 // BackendCmd represents the backend command
@@ -25,26 +22,8 @@ var BackendCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		bash_cmd := exec.Command("poetry", "run", "python", "manage.py", "runserver", fmt.Sprintf("%v", cmd.Flag("port").Value))
 		bash_cmd.Dir = "./training"
-		fmt.Println(bash_cmd.Dir)
 		fmt.Println(strings.Join(bash_cmd.Args, " "))
-		// Code below found in pty examples: https://github.com/creack/pty
-		ptmx, err := pty.Start(bash_cmd)
-		if err != nil {
-			panic(err)
-		}
-		// Make sure to close the pty at the end.
-		defer func() { _ = ptmx.Close() }() // Best effort.
-		// Set stdin in raw mode.
-		oldState, err := term.MakeRaw(int(os.Stdin.Fd()))
-		if err != nil {
-			panic(err)
-		}
-		defer func() { _ = term.Restore(int(os.Stdin.Fd()), oldState) }() // Best effort.
-
-		// Copy stdin to the pty and the pty to stdout.
-		// NOTE: The goroutine will keep reading until the next keystroke before returning.
-		go func() { io.Copy(ptmx, os.Stdin) }()
-		io.Copy(os.Stdout, ptmx)
+		pkg.ExecBashCmd(bash_cmd)
 	},
 }
